block2::ffi

Function _Block_object_assign

pub unsafe extern "C" fn _Block_object_assign(
    dest_addr: *mut c_void,
    object: *const c_void,
    flags: c_int,
)

Copy a block field or __block variable from one location to another.

Called by C compilers to clone fields inside copy helper routines, and to handle memory management of __block marked variables.