block2

Struct StackBlock

pub struct StackBlock<'f, A, R, Closure> { /* private fields */ }

An Objective-C block constructed on the stack.

This is a smart pointer that Derefs to Block.

Type parameters

The type parameters for this is a bit complex due to trait system limitations. Usually, you will not need to specify them, as the compiler should be able to infer them.

The lifetime 'f, which is the lifetime of the block.

The parameter A, which is a tuple representing the parameters to the block.

The parameter R, which is the return type of the block.

The parameter Closure, which is the contained closure type. This is usually not nameable, and you will have to use _, impl Fn() or similar.

Memory layout

The memory layout of this type is not guaranteed.

That said, it will always be safe to reintepret pointers to this as a pointer to a Block with the corresponding dyn Fn type.

Implementations

impl<'f, A, R, Closure> StackBlock<'f, A, R, Closure>

pub fn new(closure: Closure) -> Self
where A: EncodeArguments, R: EncodeReturn, Closure: IntoBlock<'f, A, R> + Clone,

Construct a StackBlock with the given closure.

Note that this requires Clone, as a C block is generally assumed to be copy-able. If you want to avoid that, put the block directly on the heap using RcBlock::new.

When the block is called, it will return the value that results from calling the closure.