Trait block2::BlockFn

pub unsafe trait BlockFn: Sealed<Self::Args, Self::Output> {
    type Args: EncodeArguments;
    type Output: EncodeReturn;
}

Types that represent closure parameters/arguments and return types in a block.

This is implemented for dyn Fn closures with up to 12 parameters, where each parameter implements EncodeArgument and the return type implements EncodeReturn.

Safety

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Required Associated Types

type Args: EncodeArguments

The parameters/arguments to the block.

type Output: EncodeReturn

The return type of the block.

Object Safety

This trait is not object safe.