block2

Trait IntoBlock

pub unsafe trait IntoBlock<'f, A, R>: Sealed<A, R>
where
    A: EncodeArguments,
    R: EncodeReturn,
{
    type Dyn: ?Sized + BlockFn<Args = A, Output = R>;
}

Types that may be converted into a block.

This is implemented for Fn closures of up to 12 parameters, where each parameter implements EncodeArgument and the return type implements EncodeReturn.

Safety

This is a sealed trait, and should not need to be implemented. Open an issue if you know a use-case where this restrition should be lifted!

Required Associated Types

type Dyn: ?Sized + BlockFn<Args = A, Output = R>

The type-erased dyn Fn(...Args) -> R + 'f.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.